Refactor relocate hiring backend logic to within academics
