+++ backend/entities/__init__.py
@@ -13,17 +13,21 @@ When adding a new entity to the application be sure to import it here. As a remi
 global to a module are available for import from other modules."""
 
 from .entity_base import EntityBase
+
 from .user_entity import UserEntity
 from .role_entity import RoleEntity
-from .room_entity import RoomEntity
 from .permission_entity import PermissionEntity
 from .user_role_table import user_role_table
+
+from .room_entity import RoomEntity
+
 from .organization_entity import OrganizationEntity
 from .event_entity import EventEntity
 from .event_registration_entity import EventRegistrationEntity
+
 from .application_entity import ApplicationEntity
 from .section_application_table import section_application_table
-from .application_review_entity import ApplicationReviewEntity
+from .academics.hiring.application_review_entity import ApplicationReviewEntity
 
 __authors__ = ["Kris Jordan"]
 __copyright__ = "Copyright 2023"
