+++ backend/main.py
@@ -18,10 +18,9 @@ from .api import (
     user,
     room,
     application,
-    hiring,
 )
 from .api.coworking import status, reservation, ambassador, operating_hours
-from .api.academics import section_member, term, course, section, my_courses
+from .api.academics import section_member, term, course, section, my_courses, hiring
 from .api.office_hours import (
     office_hours as office_hours_event,
     ticket as office_hours_ticket,
