+++ backend/services/__init__.py
@@ -6,4 +6,3 @@ from .organization import OrganizationService
 from .event import EventService
 from .exceptions import ResourceNotFoundException, UserPermissionException
 from .room import RoomService
-from .hiring import HiringService
