+++ backend/services/academics/__init__.py
@@ -2,3 +2,5 @@ from .term import TermService
 from .course import CourseService
 from .section import SectionService
 from .section_member import SectionMemberService
+
+from .hiring import HiringService
