+++ backend/test/services/fixtures.py
@@ -62,9 +62,3 @@ def event_svc_integration(session: Session, user_svc_integration: UserService):
 def room_svc(session: Session):
     """RoomService fixture."""
     return RoomService(session, PermissionService(session))
-
-
-@pytest.fixture()
-def hiring_svc(session: Session):
-    """HiringService fixture."""
-    return HiringService(session)
